# docs/openapi-mvp.yaml
openapi: 3.0.1
info:
  title: LazyPhotos API (MVP)
  version: "0.1.0"
servers:
  - url: http://localhost:5000
    description: Self-hosted baseline
paths:
...
  /albums:
    get:
      summary: List albums
      responses:
        "200":
          description: Albums
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumListResponse"
    post:
      summary: Create album
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumCreateRequest"
      responses:
        "200":
          description: Album created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
  /albums/{id}:
    put:
      summary: Update album
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumUpdateRequest"
      responses:
        "200":
          description: Album updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
  /albums/{id}/items:
    post:
      summary: Add photo to album
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumItemRequest"
      responses:
        "204":
          description: Added
    get:
      summary: List album photos
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Album photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
  /albums/{id}/items/{photoId}:
    delete:
      summary: Remove photo from album
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: photoId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    AlbumCreateRequest:
      type: object
      properties:
        name:
          type: string
      required: [name]
    AlbumUpdateRequest:
      type: object
      properties:
        name:
          type: string
        coverPhotoId:
          type: string
          format: uuid
    AlbumItemRequest:
      type: object
      properties:
        photoId:
          type: string
          format: uuid
      required: [photoId]
    AlbumListResponse:
      type: object
      properties:
        albums:
          type: array
          items:
            $ref: "#/components/schemas/Album"
      required: [albums]
...
    Photo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        storageKey:
          type: string
        hash:
          type: string
        fileName:
          type: string
        sizeBytes:
          type: integer
          format: int64
        capturedAt:
          type: string
          format: date-time
        uploadedAt:
          type: string
          format: date-time
        width:
          type: integer
        height:
          type: integer
        mimeType:
          type: string
        locationLat:
          type: number
        locationLon:
          type: number
        cameraMake:
          type: string
        cameraModel:
          type: string
        isDeleted:
          type: boolean
        updatedAt:
          type: string
          format: date-time
        thumbnails:
          type: object
          additionalProperties:
            type: string
            format: uri
      required: [id, userId, storageKey, hash, fileName, sizeBytes, uploadedAt, updatedAt]
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        coverPhotoId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isDeleted:
          type: boolean
      required: [id, userId, name, createdAt, updatedAt]